requirements.txt
pyee~=12.0.0
ucapi==0.3.1
aiohttp>=3.0.0,<4.0.0
async-timeout>=5.0.1
ssdp==1.3.0
pyamaha-yec==0.4
defusedxml==0.7.1